requirements.txt
requests
beautifulsoup4
selenium
pandas
numpy
scikit-learn
xgboost
tensorflow
torch
gym
stable-baselines3
discord.py
joblib
streamlit
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
APScheduler
